Introduce capability-utils module; treat Appium platformName and platformVersion as basic capabilities
