+++ lib/capabilityMatcher.js
@@ -6,11 +6,11 @@
 var store = require('./store'),
     models = require('./models'),
     log = require('./log'),
-    matchesCaps = require('./matcher');
+    matcher = require('./capability-utils').defaultMatcher;
 
 function matchesCapsArray(caps, nodeCapsArray) {
     return nodeCapsArray.some(function(nodeCaps) {
-        return matchesCaps(caps, nodeCaps);
+        return matcher(caps, nodeCaps);
     });
 }
